feat: make exception error

+++ client/src/views/Login.vue
@@ -15,7 +15,7 @@
         </div>
         <button @click="login" class="bg-[#EB6648] text-white w-full rounded-md py-1.5">submit</button>
         <p class="text-md text-gray-400 text-center mt-20">
-          don't have an account? <a @click="this.$router.push('/login')" class="text-black/70 cursor-pointer block">Register now ✨</a>
+          don't have an account? <a @click="this.$router.push('/register')" class="text-black/70 cursor-pointer block">Register now ✨</a>
         </p>
       </div>
     </div>
@@ -25,29 +25,75 @@
 <script>
 import Layout from "../components/Layout.vue";
 import axios from "axios";
+import useValidate from "@vuelidate/core";
+import { required, email, minLength } from "@vuelidate/validators";
 export default {
   components: {
     Layout,
   },
   data() {
     return {
+      v$: useValidate(),
       data: {
         email: "",
         password: "",
       },
     };
   },
-  onMounted() {},
+  validations() {
+    return {
+      data: {
+        email: {
+          required,
+          email,
+        },
+        password: {
+          required,
+          minLength: minLength(6),
+        },
+      },
+    };
+  },
   methods: {
+    async showAlert(type, text) {
+      // Use sweetalert2
+      const Toast = await this.$swal.mixin({
+        toast: true,
+        position: "top-end",
+        showConfirmButton: false,
+        timer: 3000,
+        timerProgressBar: true,
+        didOpen: (toast) => {
+          toast.addEventListener("mouseenter", this.$swal.stopTimer);
+          toast.addEventListener("mouseleave", this.$swal.resumeTimer);
+        },
+      });
+
+      Toast.fire({
+        icon: type,
+        title: text,
+      });
+    },
     async login() {
       try {
+        const result = await this.v$.$validate();
+
+        if (!result) {
+          // notify user form is invalid
+          throw new Error("Invalid data");
+        }
+
         const res = await axios.post("http://localhost:8080/api/user/login", this.data);
-        console.log(res.data);
         localStorage.setItem("user", JSON.stringify(res.data.user));
         localStorage.setItem("token", res.data.accessToken);
         this.$router.push("/");
+        this.showAlert("success", "Successfully registered");
       } catch (error) {
-        console.log(error.response.data);
+        if (error?.response?.data?.message) {
+          this.showAlert("error", error?.response?.data?.message);
+        } else {
+          this.showAlert("error", error);
+        }
       }
     },
   },
